src/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "advanced-rag"
version = "0.1.0"
description = "Advanced RAG using langchain and fastapi"
authors = [
    {name = "Sameeta Kumavat",email = "sameetakumavat@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "bcrypt (==4.0.1)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "langgraph (>=0.3.31,<0.4.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "langchain-groq (>=0.3.2,<0.4.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "wikipedia (>=1.4.0,<2.0.0)",
    "streamlit (>=1.45.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
